### 2025-08-21 #8840 @editorialbot
[PRE REVIEW]: NCALab: A Framework for Experimentation with Neural Cellular Automata
**Submitting author:** @henrykrumb (<a href="http://orcid.org/0000-0001-8189-4752">Henry John Krumb</a>)
**Repository:** https://github.com/MECLabTUDA/NCALab
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.2
