### 2025-09-19 72ef47f
add cifar10 example task
