### 2025-09-22 c5bed75
bump keras dependency (due to security alert)
