### 2025-09-23 11f49f2
classification: parameterize average pooling size
