### 2025-09-23 1ca6649
visualization: show horizontal bars with softmax probs
