### 2025-09-23 29b6d0a
add scipy type stubs
