+++ pyproject.toml
@@ -191,6 +191,7 @@ dependencies = [
     "ruff>=0.12.4",
     "torchmetrics>=1.7.4",
     "segmentation-models-pytorch>=0.5.0",
+    "scipy-stubs>=1.16.2.0",
 ]
 
 
