### 2025-09-23 90b79fc
append recent changes to changelog
