### 2025-09-24 451fa6e
cifar10 example: add more augmentations, add evaluation script
