+++ tasks/class_cifar10/train_class_cifar10.py
@@ -53,6 +53,10 @@ def train_class_cifar10(
             v2.ToImage(),
             v2.ToDtype(torch.float, scale=True),
             v2.ConvertImageDtype(dtype=torch.float32),
+            transforms.ColorJitter(
+                brightness=0.1, contrast=0.1, saturation=0.1, hue=0.1
+            ),
+            transforms.RandomPerspective(distortion_scale=0.5, p=0.5),
             transforms.RandomHorizontalFlip(),
             transforms.RandomVerticalFlip(),
             transforms.Normalize((0.5, 0.5, 0.5), (0.5, 0.5, 0.5)),
