### 2025-09-25 fe9ca3d
fix: scheduler step after each epoch
