### 2025-09-25 f000a13
remove dropout in classification NCA head
