### 2025-09-25 1ae62d0
separate validation and training transforms in cifar10 example
