### 2025-10-01 5b6e70c
example tasks: organize imports and ignore E402 for ncalab imports
