### 2025-10-02 22f6a88
use Prediction class in metrics computation, split head/prediction model from basicNCA
