+++ ncalab/models/__init__.py
@@ -1,4 +1,5 @@
 from .basicNCA import BasicNCAModel  # noqa: F401
+from .basicNCArule import BasicNCARule  # noqa: F401
 from .cascadeNCA import CascadeNCA  # noqa: F401
 from .classificationNCA import ClassificationNCAModel  # noqa: F401
 from .depthNCA import DepthNCAModel  # noqa: F401
