+++ ncalab/models/basicNCA.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
-from typing import Any, Dict, Optional, Tuple, List
+
+from typing import Any, Dict, List, Optional, Tuple
 
 import numpy as np
-
 import torch  # type: ignore[import-untyped]
 import torch.nn as nn  # type: ignore[import-untyped]
 import torch.nn.functional as F  # type: ignore[import-untyped]
@@ -11,6 +11,8 @@ from ..autostepper import AutoStepper
 from ..prediction import Prediction
 from ..utils import pad_input
 from ..visualization import Visual
+from .basicNCArule import BasicNCARule
+from .basicNCAhead import BasicNCAHead
 
 
 class BasicNCAModel(nn.Module):
@@ -37,22 +39,23 @@ class BasicNCAModel(nn.Module):
         pad_noise: bool = False,
         autostepper: Optional[AutoStepper] = None,
         use_temporal_encoding: bool = False,
+        rule_type: type[BasicNCARule] = BasicNCARule,
     ):
         """
-        :param device [device]: Pytorch device descriptor.
-        :param num_image_channels [int]: Number of channels reserved for input image.
-        :param num_hidden_channels [int]: Number of hidden channels (communication channels).
-        :param num_output_channels [int]: Number of output channels.
-        :param fire_rate [float]: Fire rate for stochastic weight update. Defaults to 0.5.
-        :param hidden_size [int]: Number of neurons in hidden layer. Defaults to 128.
-        :param use_alive_mask [bool]: Whether to use alive masking (channel 3) during training. Defaults to False.
-        :param immutable_image_channels [bool]: If image channels should be fixed during inference, which is the case for most segmentation or classification problems. Defaults to True.
-        :param num_learned_filters [int]: Number of learned filters. If zero, use two sobel filters instead. Defaults to 2.
-        :param filter_padding [str]: Padding type to use. Might affect reliance on spatial cues. Defaults to "circular".
-        :param use_laplace [bool]: Whether to use Laplace filter (only if num_learned_filters == 0)
-        :param kernel_size [int]: Filter kernel size (only for learned filters)
-        :param pad_noise [bool]: Whether to pad input image tensor with noise in hidden / output channels
-        :param autostepper [Optional[AutoStepper]]: AutoStepper object to select number of time steps based on activity
+        :param device: Pytorch device descriptor.
+        :param num_image_channels: Number of channels reserved for input image.
+        :param num_hidden_channels: Number of hidden channels (communication channels).
+        :param num_output_channels: Number of output channels.
+        :param fire_rate: Fire rate for stochastic weight update. Defaults to 0.5.
+        :param hidden_size: Number of neurons in hidden layer. Defaults to 128.
+        :param use_alive_mask: Whether to use alive masking (channel 3) during training. Defaults to False.
+        :param immutable_image_channels: If image channels should be fixed during inference, which is the case for most segmentation or classification problems. Defaults to True.
+        :param num_learned_filters: Number of learned filters. If zero, use two sobel filters instead. Defaults to 2.
+        :param filter_padding: Padding type to use. Might affect reliance on spatial cues. Defaults to "circular".
+        :param use_laplace: Whether to use Laplace filter (only if num_learned_filters == 0)
+        :param kernel_size: Filter kernel size (only for learned filters)
+        :param pad_noise: Whether to pad input image tensor with noise in hidden / output channels
+        :param autostepper: AutoStepper object to select number of time steps based on activity
         """
         super(BasicNCAModel, self).__init__()
 
@@ -83,41 +86,24 @@ class BasicNCAModel(nn.Module):
         self._define_filters(num_learned_filters)
 
[…]
