### 2025-10-02 b11280c
fix use of api in depth estimation model
