### 2025-10-07 d537252
apply several fixes in classification model logic
