### 2025-10-07 af873b6
fix: don't use head for MNIST classification
