### 2025-10-07 1366b8d
use cosine annealing schedule
