### 2025-10-07 881e19f
create shell script for training all example tasks
