### 2025-10-07 39fd3d8
fix cifar10 evaluation: apply same normalization
