+++ tasks/class_cifar10/eval_class_cifar10.py
@@ -35,7 +35,7 @@ T = transforms.Compose(
         v2.ToImage(),
         v2.ToDtype(torch.float, scale=True),
         v2.ConvertImageDtype(dtype=torch.float32),
-        transforms.Normalize((0.5, 0.5, 0.5), (0.5, 0.5, 0.5)),
+        transforms.Normalize((0.4914, 0.4822, 0.4465), (0.2023, 0.1994, 0.2010)),
     ]
 )
 
