### 2025-10-07 7675037
show ncalab banner in medmnist examples
