+++ CHANGELOG.md
@@ -1,7 +1,9 @@
-# v0.3.3
+# v0.4.0
 * Parameterize average pooling size in classification with MLP head
 * Let user supply class label names to classifier
 * Classification visual in tensorboard: Show bar plot of softmax
+* Separation of rule and head model in BasicNCAModel for improved modularity
+* Classification: Average pooling between NCA backbone and classification head
 
 # v0.3.2
 * Fix use of max() reduction in uncertainty estimator
