### 2025-11-28 f5e02b2
move training/inference steps to models, refactor perception module
