### 2025-11-28 33f7c66
add keras==3.12 to fix security warning
