+++ pyproject.toml
@@ -49,7 +49,7 @@ dependencies = [
     "jedi==0.19.2",
     "jinja2==3.1.6",
     "joblib==1.4.2",
-    "keras==3.11.3",
+    "keras==3.12.0",
     "keras-applications==1.0.8",
     "kiwisolver==1.4.8",
     "lazy-loader==0.4",
