### 2025-12-01 d98ae4d
describe NCA model structure in more detail
