+++ ncalab/models/basicNCA/__init__.py
@@ -1,0 +1,4 @@
+from .basicNCA import BasicNCAModel  # noqa: F401
+from .basicNCAhead import BasicNCAHead  # noqa: F401
+from .basicNCAperception import BasicNCAPerception  # noqa: F401
+from .basicNCArule import BasicNCARule  # noqa: F401
