+++ ncalab/models/applications/__init__.py
@@ -1,0 +1,4 @@
+from .classification import *  # noqa: F401,F403
+from .depth import *  # noqa: F401,F403
+from .growing import *  # noqa: F401,F403
+from .segmentation import *  # noqa: F401,F403
