+++ ncalab/models/applications/segmentation/__init__.py
@@ -1,0 +1,1 @@
+from .segmentationNCA import SegmentationNCAModel  # noqa: F401
