+++ ncalab/models/applications/classification/__init__.py
@@ -1,0 +1,2 @@
+from .classificationNCA import ClassificationNCAModel  # noqa: F401
+from .classificationNCAhead import ClassificationNCAHead  # noqa: F401
