+++ ncalab/models/applications/depth/__init__.py
@@ -1,0 +1,1 @@
+from .depthNCA import DepthNCAModel  # noqa: F401
