+++ ncalab/models/wrappers/__init__.py
@@ -1,0 +1,1 @@
+from .cascadeNCA import CascadeNCA  # noqa: F401
