+++ pyproject.toml
@@ -34,7 +34,6 @@ dependencies = [
     "filelock==3.18.0",
     "fire==0.7.0",
     "flatbuffers==25.2.10",
-    "fonttools==4.57.0",
     "fsspec==2025.3.2",
     "gast==0.6.0",
     "google-pasta==0.2.0",
@@ -133,7 +132,7 @@ dependencies = [
     "tzdata==2025.2",
     "urllib3==2.5.0",
     "wcwidth==0.2.13",
-    "werkzeug==3.1.3",
+    "werkzeug==3.1.4",
     "wheel==0.46.1",
     "wrapt==1.17.2",
     "zipp==3.21.0",
@@ -192,6 +191,7 @@ dependencies = [
     "torchmetrics>=1.7.4",
     "segmentation-models-pytorch>=0.5.0",
     "scipy-stubs>=1.16.2.0",
+    "fonttools>=4.57.0",
 ]
 
 
