### 2025-12-02 454773a
upgrade pytorch to 2.8.0
