+++ pyproject.toml
@@ -25,7 +25,6 @@ dependencies = [
     "debugpy==1.8.14",
     "decorator==5.2.1",
     "numpy==2.1.3",
-    "torch==2.7.1",
     "tqdm==4.67.1",
     "emoji==2.14.1",
     "etils==1.12.2",
@@ -121,7 +120,6 @@ dependencies = [
     "tifffile==2025.3.30",
     "tokenizers==0.21.1",
     "torch-tb-profiler==0.4.3",
-    "torchvision==0.22.1",
     "tornado==6.5",
     "traitlets==5.14.3",
     "transformers==4.53",
@@ -150,20 +148,6 @@ dependencies = [
     "jaraco-functools==4.1.0",
     "more-itertools==10.7.0",
     "multidict==6.4.3",
-    "nvidia-cublas-cu12==12.6.4.1",
-    "nvidia-cuda-cupti-cu12==12.6.80",
-    "nvidia-cuda-nvrtc-cu12==12.6.77",
-    "nvidia-cuda-runtime-cu12==12.6.77",
-    "nvidia-cudnn-cu12==9.5.1.17",
-    "nvidia-cufft-cu12==11.3.0.4",
-    "nvidia-cufile-cu12==1.11.1.6",
-    "nvidia-curand-cu12==10.3.7.77",
-    "nvidia-cusolver-cu12==11.7.1.2",
-    "nvidia-cusparse-cu12==12.5.4.2",
-    "nvidia-cusparselt-cu12==0.6.3",
-    "nvidia-nccl-cu12==2.26.2",
-    "nvidia-nvjitlink-cu12==12.6.85",
-    "nvidia-nvtx-cu12==12.6.77",
     "oauthlib==3.2.2",
     "pip-tools==7.4.1",
     "pip-upgrader==1.4.15",
@@ -192,6 +176,7 @@ dependencies = [
     "segmentation-models-pytorch>=0.5.0",
     "scipy-stubs>=1.16.2.0",
     "fonttools>=4.57.0",
+    "torch==2.8.0",
 ]
 
 
