### 2025-12-02 db3bcbe
create script for dry-running all example tasks
