+++ tasks/class_medmnist/train_class_bloodmnist.py
@@ -38,6 +38,8 @@ def train_class_bloodmnist(
     hidden_channels: int,
     gpu: bool,
     gpu_index: int,
+    max_epochs: int,
+    dry: bool,
 ):
     print_NCALab_banner()
 
@@ -48,7 +50,7 @@ def train_class_bloodmnist(
     comment += f"_AM_{alive_mask}"
     comment += f"_TE_{use_temporal_encoding}"
 
-    writer = SummaryWriter(comment=comment)
+    writer = SummaryWriter(comment=comment) if not dry else None
 
     device = get_compute_device(f"cuda:{gpu_index}" if gpu else "cpu")
 
@@ -84,22 +86,23 @@ def train_class_bloodmnist(
         pad_noise=pad_noise,
         use_temporal_encoding=use_temporal_encoding,
         class_names=list(INFO["bloodmnist"]["label"].values()),
+        training_timesteps=32,
+        inference_timesteps=32,
     )
     trainer = BasicNCATrainer(
         nca,
-        WEIGHTS_PATH / "classification_bloodmnist",
+        WEIGHTS_PATH / "classification_bloodmnist" if not dry else None,
         batch_repeat=2,
-        max_epochs=40,
+        max_epochs=max_epochs,
         gradient_clipping=gradient_clipping,
-        steps_range=(32, 33),
-        steps_validation=32,
     )
     trainer.train(
         loader_train,
         loader_val,
         summary_writer=writer,
     )
-    writer.close()
+    if writer is not None:
+        writer.close()
 
 
 @click.command()
@@ -111,12 +114,18 @@ def train_class_bloodmnist(
 @click.option(
     "--gpu-index", type=int, default=0, help="Index of GPU to use, if --gpu in use."
 )
-def main(batch_size, hidden_channels, gpu: bool, gpu_index: int):
+@click.option("--max-epochs", "-E", type=int, default=40)
+@click.option("--dry", "-D", is_flag=True)
+def main(
+    batch_size, hidden_channels, gpu: bool, gpu_index: int, max_epochs: int, dry: bool
+):
     train_class_bloodmnist(
         batch_size=batch_size,
         hidden_channels=hidden_channels,
         gpu=gpu,
         gpu_index=gpu_index,
+        max_epochs=max_epochs,
+        dry=dry,
     )
 
 
