### 2025-12-02 745697e
update version in project config
