### 2026-01-18 #8840 @sepandhaghighi
> [@blsqr](https://github.com/blsqr) , [@sepandhaghighi](https://github.com/sepandhaghighi) , [@neuromusic](https://github.com/neuromusic) , [@schnorr](https://github.com/schnorr) - Would any of you be interested in reviewing this paper for JOSS?

Hi @kanishkan91
I am available to review this paper 💯
