### 2026-01-28 #8840 @arfon
Apologies @henrykrumb @kanishkan91, we missed one section from the list above. In addition to the sections already mentioned, JOSS papers now also require:

Apologies again for the oversight.
