### 2026-02-13 #8840 @schnorr
> [@blsqr](https://github.com/blsqr) , [@sepandhaghighi](https://github.com/sepandhaghighi) , [@neuromusic](https://github.com/neuromusic) , [@schnorr](https://github.com/schnorr) - Would any of you be interested in reviewing this paper for JOSS?

Dear @kanishkan91, thanks for the invitation to review! I'll be happy to help.

Should we wait for the author to modify the content of the submission, as instructed by @arfon above?
