### 2026-02-23 597c807
update whitepaper, adding sections on use of genAI and research impact
