+++ docs/source/configuration.rst
@@ -1,0 +1,7 @@
+.. title:: NCALab Configuration
+.. author:: Henry Krumb
+.. date:: 2025-12-03
+
+=============
+Configuration
+=============
