+++ docs/source/training.rst
@@ -1,0 +1,7 @@
+.. title:: Introduction to NCALab
+.. author:: Henry Krumb
+.. date:: 2025-12-01
+
+=======================
+Training NCAs in NCALab
+=======================
