### 2026-02-23 bd8b917
bump dependencies that caused dependabot alerts
