+++ pyproject.toml
@@ -47,7 +47,7 @@ dependencies = [
     "jedi==0.19.2",
     "jinja2==3.1.6",
     "joblib==1.4.2",
-    "keras==3.12.0",
+    "keras==3.12.1",
     "keras-applications==1.0.8",
     "kiwisolver==1.4.8",
     "lazy-loader==0.4",
@@ -128,7 +128,7 @@ dependencies = [
     "types-tqdm==4.67.0.20250417",
     "typing-extensions==4.13.2",
     "tzdata==2025.2",
-    "urllib3==2.5.0",
+    "urllib3==2.6.3",
     "wcwidth==0.2.13",
     "werkzeug==3.1.4",
     "wheel==0.46.1",
@@ -153,7 +153,7 @@ dependencies = [
     "pip-upgrader==1.4.15",
     "propcache==0.3.1",
     "proto-plus==1.26.1",
-    "pyasn1==0.6.1",
+    "pyasn1==0.6.2",
     "pyasn1-modules==0.4.2",
     "pyproject-hooks==1.2.0",
     "requests-oauthlib==2.0.0",
