+++ pyproject.toml
@@ -30,7 +30,7 @@ dependencies = [
     "etils==1.12.2",
     "eval-type-backport==0.2.2",
     "executing==2.2.0",
-    "filelock==3.18.0",
+    "filelock==3.20.1",
     "fire==0.7.0",
     "flatbuffers==25.2.10",
     "fsspec==2025.3.2",
@@ -130,13 +130,13 @@ dependencies = [
     "tzdata==2025.2",
     "urllib3==2.6.3",
     "wcwidth==0.2.13",
-    "werkzeug==3.1.4",
-    "wheel==0.46.1",
+    "werkzeug==3.1.5",
+    "wheel==0.46.2",
     "wrapt==1.17.2",
     "zipp==3.21.0",
     "sphinx-autoapi>=3.6.0",
     "aiohappyeyeballs==2.6.1",
-    "aiohttp==3.12.14",
+    "aiohttp==3.13.3",
     "attrs==25.3.0",
     "build==1.2.2.post1",
     "cachetools==5.5.2",
@@ -166,7 +166,7 @@ dependencies = [
     "identify==2.6.10",
     "nodeenv==1.9.1",
     "pre-commit==4.2.0",
-    "virtualenv==20.30.0",
+    "virtualenv==20.36.1",
     "sphinx==8.1.3",
     "sphinx-rtd-theme>=3.0.2",
     "plyfile>=1.1",
