+++ pyproject.toml
@@ -30,7 +30,7 @@ dependencies = [
     "etils==1.12.2",
     "eval-type-backport==0.2.2",
     "executing==2.2.0",
-    "filelock==3.20.1",
+    "filelock==3.20.3",
     "fire==0.7.0",
     "flatbuffers==25.2.10",
     "fsspec==2025.3.2",
@@ -77,12 +77,12 @@ dependencies = [
     "pandas-stubs==2.2.3.250308",
     "parso==0.8.4",
     "pexpect==4.9.0",
-    "pillow==11.3.0",
+    "pillow==12.1.1",
     "platformdirs==4.3.7",
     "pluggy==1.5.0",
     "pretrainedmodels==0.7.4",
     "prompt-toolkit==3.0.51",
-    "protobuf==4.25.8",
+    "protobuf==5.29.6",
     "psutil==7.0.0",
     "ptyprocess==0.7.0",
     "pure-eval==0.2.3",
@@ -113,7 +113,6 @@ dependencies = [
     "sympy==1.14.0",
     "tensorboard==2.19.0",
     "tensorboard-data-server==0.7.2",
-    "tensorboard-plugin-profile==2.19.4",
     "tensorflow==2.19.0",
     "termcolor==3.1.0",
     "threadpoolctl==3.6.0",
@@ -130,7 +129,7 @@ dependencies = [
     "tzdata==2025.2",
     "urllib3==2.6.3",
     "wcwidth==0.2.13",
-    "werkzeug==3.1.5",
+    "werkzeug==3.1.6",
     "wheel==0.46.2",
     "wrapt==1.17.2",
     "zipp==3.21.0",
@@ -177,6 +176,7 @@ dependencies = [
     "scipy-stubs>=1.16.2.0",
     "fonttools>=4.57.0",
     "torch==2.8.0",
+    "tensorboard-plugin-profile>=2.21.6",
 ]
 
 
