### 2026-02-24 05ebbfa
comment NCA model code, fix potential issue with temporal encoding
