+++ ncalab/models/README.md
@@ -1,0 +1,19 @@
+# NCA Models
+
+
+## applications/
+
+Contains model definitions for some common downstream tasks.
+
+**classification**: NCA model for multiclass or binary image segmentation.
+
+**depth**: NCA model for monocular depth estimation.
+
+**growing**: "Growing NCA", similar to the idea of Mordvintsev et al.
+
+**segmentation**: NCA model for pixel-wise image segmentation. Currently supports binary segmentation only.
+
+
+## basicNCA/
+
+## wrappers/
