### 2026-03-02 a244f45
paper: add state of the field
