### 2026-03-10 15996d5
improve prediction animations
