### 2026-03-11 2a97850
create hidden channel animation
