+++ README.md
@@ -13,6 +13,7 @@ For more information on NCAs, check out our curated [Awesome List](https://githu
 
 ![Animation of a growing lizard emoji](artwork/growing_emoji.gif)
 ![Animation of gastro-intestinal polyp segmentation using NCA](artwork/segmentation_kvasir_seg.gif)
+![Animation of hidden channels of polyp segmentation](artwork/segmentation_kvasir_seg_hidden.gif)
 
 ## Neural Cellular Automata
 
