### 2026-03-11 08f575a
add cifar10 figure to README
