### 2026-03-11 9069a1d
add show_input option to animator, showing the input image next to prediction
