### 2026-03-11 c6df3d8
generate gif in cifar10 evaluation
