### 2026-03-11 e447735
generate gif for dermamnist prediction in feature space
