+++ README.md
@@ -15,6 +15,7 @@ For more information on NCAs, check out our curated [Awesome List](https://githu
 ![Animation of gastro-intestinal polyp segmentation using NCA](artwork/segmentation_kvasir_seg.gif)
 ![Animation of hidden channels of polyp segmentation](artwork/segmentation_kvasir_seg_hidden.gif)
 ![Animation of a prediction on CIFAR-10 in feature space](artwork/classification_cifar10.gif)
+![Animation of a prediction on DermaMNIST in feature space](artwork/classification_dermamnist.gif)
 
 ## Neural Cellular Automata
 
