### 2026-03-12 1b20409
tasks: generate animated figures
