+++ tasks/class_medmnist/eval_class_pathmnist.py
@@ -11,6 +11,7 @@ from torchvision import transforms  # type: ignore[import-untyped]
 from torchvision.transforms import v2  # type: ignore[import-untyped]
 from tqdm import tqdm
 from train_class_pathmnist import (
+    TASK_PATH,
     WEIGHTS_PATH,
     alive_mask,
     default_hidden_channels,
@@ -22,7 +23,10 @@ from train_class_pathmnist import (
 root_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), "../.."))
 sys.path.append(root_dir)
 
-from ncalab import ClassificationNCAModel, get_compute_device  # noqa: E402
+from ncalab import Animator, ClassificationNCAModel, get_compute_device  # noqa: E402
+
+FIGURE_PATH = TASK_PATH / "figures"
+FIGURE_PATH.mkdir(exist_ok=True)
 
 T = transforms.Compose(
     [
@@ -51,12 +55,14 @@ def eval_selfclass_pathmnist(
         device,
         num_image_channels=3,
         num_hidden_channels=hidden_channels,
-        hidden_size=128,
         num_classes=num_classes,
         use_alive_mask=alive_mask,
         fire_rate=fire_rate,
-        use_temporal_encoding=use_temporal_encoding,
         pad_noise=pad_noise,
+        use_temporal_encoding=use_temporal_encoding,
+        class_names=list(INFO["pathmnist"]["label"].values()),
+        training_timesteps=32,
+        inference_timesteps=32,
     )
     nca.load_state_dict(
         torch.load(
@@ -88,9 +94,9 @@ def eval_selfclass_pathmnist(
     for sample in tqdm(iter(loader_test)):
         x, y = sample
         x = x.float().to(device)
-        steps = 32
-        y_prob = nca.classify(x, steps, reduce=False)
         y = y.squeeze().to(device)
+        steps = 32
+        y_prob = nca.classify(x, steps=steps, reduce=False)
 
         macro_acc.update(y_prob, y)
         micro_acc.update(y_prob, y)
@@ -100,6 +106,12 @@ def eval_selfclass_pathmnist(
     micro_acc_ = micro_acc.compute().item()
     macro_auc_ = macro_auc.compute().item()
     micro_f1_ = micro_f1.compute().item()
+
+    seed = next(iter(loader_test))[0].to(device)
+    out_path = FIGURE_PATH / "classification_pathmnist.gif"
+    animator = Animator(nca, seed, interval=100, steps=32, hidden=True, show_input=True)
+    animator.save(out_path)
+    print(f"Animation saved to: {out_path}")
 
     print()
     print(
