+++ pyproject.toml
@@ -52,7 +52,7 @@ dependencies = [
     "kiwisolver==1.4.8",
     "lazy-loader==0.4",
     "libclang==18.1.1",
-    "markdown==3.8",
+    "markdown==3.8.1",
     "markdown-it-py==3.0.0",
     "markupsafe==3.0.2",
     "matplotlib==3.10.1",
@@ -119,7 +119,6 @@ dependencies = [
     "tifffile==2025.3.30",
     "tokenizers==0.21.1",
     "torch-tb-profiler==0.4.3",
-    "tornado==6.5",
     "traitlets==5.14.3",
     "transformers==4.53",
     "types-pytz==2025.2.0.20250326",
@@ -177,6 +176,7 @@ dependencies = [
     "fonttools>=4.57.0",
     "torch==2.8.0",
     "tensorboard-plugin-profile>=2.21.6",
+    "tornado>=6.5.5",
 ]
 
 
