### 2026-03-29 ed2ea64
rename 'basicNCAmodel' and associated models to 'abstractNCAmodel'
