+++ ncalab/models/wrappers/cascadeNCA.py
@@ -1,11 +1,11 @@
 import copy
-from typing import Dict, List, Optional, Tuple
+from typing import List, Optional, Tuple
 
 import numpy as np
 import torch  # type: ignore[import-untyped]
 import torch.nn as nn  # type: ignore[import-untyped]
 
-from ncalab.models.basicNCA import BasicNCAModel
+from ncalab.models.basicNCA import AbstractNCAModel
 from ncalab.prediction import Prediction
 from ncalab.utils import unwrap
 
@@ -25,7 +25,8 @@ def upscale(image: torch.Tensor, scale: float, mode: str = "nearest") -> torch.T
     :rtype: torch.Tensor
     """
     assert scale >= 1.0
-    return nn.Upsample(scale_factor=scale, mode=mode)(image)
+    x = nn.Upsample(scale_factor=scale, mode=mode)(image)
+    return x
 
 
 def downscale(
@@ -44,10 +45,11 @@ def downscale(
     :rtype: torch.Tensor
     """
     assert scale >= 1.0
-    return nn.Upsample(scale_factor=1.0 / scale, mode=mode, align_corners=True)(image)
+    x = nn.Upsample(scale_factor=1.0 / scale, mode=mode, align_corners=True)(image)
+    return x
 
 
-class CascadeNCA(BasicNCAModel):
+class CascadeNCA(AbstractNCAModel):
     """
     Chain multiple instances of the same NCA model, operating at different
     image scales.
@@ -62,14 +64,14 @@ class CascadeNCA(BasicNCAModel):
 
     def __init__(
         self,
-        wrapped: BasicNCAModel,
+        wrapped: AbstractNCAModel,
         scales: List[int],
         steps: List[int],
         single_model: bool = True,
     ):
         """
-        :param wrapped: Backbone model based on BasicNCAModel.
-        :type wrapped: ncalab.BasicNCAModel
+        :param wrapped: Backbone model based on AbstractNCAModel.
+        :type wrapped: ncalab.AbstractNCAModel
         :param scales: List of scales to operate at, e.g. [4, 2, 1].
         :type scales: List[int]
         :param steps: List of number of NCA inference time steps.
@@ -105,6 +107,10 @@ class CascadeNCA(BasicNCAModel):
             self.num_classes = wrapped.num_classes
         if hasattr(wrapped, "avg_pool_size"):
             self.avg_pool_size = wrapped.avg_pool_size
+        if hasattr(wrapped, "class_names"):
+            self.class_names = wrapped.class_names
+
+        self.metrics = wrapped.metrics
 
         self.wrapped = wrapped
         assert len(scales) == len(steps)
@@ -116,7 +122,7 @@ class CascadeNCA(BasicNCAModel):
         self.steps = steps
 
         self.single_model = single_model
-        self.models: List[BasicNCAModel]
+        self.models: List[AbstractNCAModel]
         if single_model:
             self.models = [wrapped for _ in scales]
         else:
@@ -135,8 +141,8 @@ class CascadeNCA(BasicNCAModel):
         assert len(self.scales) > 0
         assert len(self.models) > 0
         assert len(self.steps) > 0
-        prediction = None
         x_scaled = downscale(x, self.scales[0])
+        total_steps = 0
         for i, (model, scale, scale_steps) in enumerate(
             zip(self.models, self.scales, self.steps)
         ):
@@ -145,20 +151,38 @@ class CascadeNCA(BasicNCAModel):
             )
             if steps <= 0:
                 steps = 1
-            prediction = model(x_scaled, steps=steps)
-            if i < len(self.scales) - 1:
-                x_scaled = upscale(prediction.output_image, scale / self.scales[i + 1])
-                # replace input with downscaled variant of original image
[…]
