+++ ncalab/models/basicNCA/__init__.py
@@ -1,4 +1,5 @@
-from .basicNCA import BasicNCAModel  # noqa: F401
-from .basicNCAhead import BasicNCAHead  # noqa: F401
+from .abstractNCA import AbstractNCAModel  # noqa: F401
+from .abstractNCAhead import AbstractNCAHead  # noqa: F401
 from .basicNCAperception import BasicNCAPerception  # noqa: F401
-from .basicNCArule import BasicNCARule  # noqa: F401
+from .abstractNCArule import AbstractNCARule  # noqa: F401
+from .mlpNCArule import MLPNCARule  # noqa: F401
