### 2026-03-29 ad926ed
growing emoji task: use hidden state loss
