+++ ncalab/training/__init__.py
@@ -1,4 +1,5 @@
 from .earlystopping import *  # noqa: F401,F403
 from .kfold import *  # noqa: F401,F403
+from .pool import Pool  # noqa: F401,F403
 from .trainer import BasicNCATrainer # noqa: F401,F403
 from .traininghistory import TrainingHistory  # noqa: F401,F403
