+++ tasks/growing_emoji/train_growing_emoji.py
@@ -58,6 +58,7 @@ def train_growing_emoji(
         num_image_channels=4,
         num_hidden_channels=hidden_channels,
         use_alive_mask=True,
+        lambda_hidden=1e-2,
     )
 
     # Create dataset containing a single growing emoji
