### 2026-03-29 53810a7
use hidden state loss and new validation routine in classification tasks
