+++ tasks/class_medmnist/eval_class_bloodmnist.py
@@ -4,12 +4,7 @@ import sys
 
 import click
 import torch  # type: ignore[import-untyped]
-import torchmetrics
-import torchmetrics.classification
 from medmnist import INFO, BloodMNIST  # type: ignore[import-untyped]
-from torchvision import transforms  # type: ignore[import-untyped]
-from torchvision.transforms import v2  # type: ignore[import-untyped]
-from tqdm import tqdm
 from train_class_bloodmnist import (
     WEIGHTS_PATH,
     alive_mask,
@@ -17,22 +12,20 @@ from train_class_bloodmnist import (
     fire_rate,
     pad_noise,
     use_temporal_encoding,
+    TASK_PATH,
+    T_val
 )
+import pprint
 
 root_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), "../.."))
 sys.path.append(root_dir)
 
 
-from ncalab import ClassificationNCAModel, get_compute_device  # noqa: E402
+from ncalab import fix_random_seed, Animator, ClassificationNCAModel, get_compute_device  # noqa: E402
 
-T = transforms.Compose(
-    [
-        v2.ToImage(),
-        v2.ToDtype(torch.float, scale=True),
-        v2.ConvertImageDtype(dtype=torch.float32),
-        transforms.Normalize((0.5,), (0.225,)),
-    ]
-)
+FIGURE_PATH = TASK_PATH / "figures"
+FIGURE_PATH.mkdir(exist_ok=True)
+
 
 
 def eval_selfclass_bloodmnist(
@@ -40,11 +33,13 @@ def eval_selfclass_bloodmnist(
     gpu,
     gpu_index,
 ):
+    fix_random_seed()
+
     device = get_compute_device(f"cuda:{gpu_index}" if gpu else "cpu")
 
-    dataset_test = BloodMNIST(split="test", download=True, transform=T)
+    dataset_test = BloodMNIST(split="test", download=True, transform=T_val)
     loader_test = torch.utils.data.DataLoader(
-        dataset_test, shuffle=False, batch_size=32
+        dataset_test, shuffle=True, batch_size=8
     )
 
     num_classes = len(INFO["bloodmnist"]["label"])
@@ -52,12 +47,16 @@ def eval_selfclass_bloodmnist(
         device,
         num_image_channels=3,
         num_hidden_channels=hidden_channels,
-        hidden_size=128,
         num_classes=num_classes,
         use_alive_mask=alive_mask,
         fire_rate=fire_rate,
-        use_temporal_encoding=use_temporal_encoding,
         pad_noise=pad_noise,
+        use_temporal_encoding=use_temporal_encoding,
+        class_names=list(INFO["bloodmnist"]["label"].values()),
+        training_timesteps=32,
+        inference_timesteps=32,
+        use_classifier=True,
+        lambda_hidden=1e-2,
     )
     nca.load_state_dict(
         torch.load(
@@ -70,44 +69,16 @@ def eval_selfclass_bloodmnist(
     print(f"Trainable parameters: {params}")
     print(f"That is {4 * params / 1000} kB")
 
-    nca = nca.to(device)
-    nca.eval()
+    metrics, _ = nca.validate(loader_test)
 
-    macro_acc = torchmetrics.classification.MulticlassAccuracy(
-        average="macro", num_classes=num_classes
-    ).to(device)
-    micro_acc = torchmetrics.classification.MulticlassAccuracy(
-        average="micro", num_classes=num_classes
-    ).to(device)
-    macro_auc = torchmetrics.classification.MulticlassAUROC(
-        average="macro",
-        num_classes=num_classes,
-    ).to(device)
[…]
