### 2026-03-29 bb39e70
fix variable names in mlp rule
