+++ ncalab/models/__init__.py
@@ -1,5 +1,3 @@
 from .basicNCA import *  # noqa: F401,F403
-from .extremeNCA import *  # noqa: F401,F403
-from .rbfNCA import *  # noqa: F401,F403
 from .wrappers import *  # noqa: F401,F403
 from .applications import *  # noqa: F401,F403
