+++ pyproject.toml
@@ -180,6 +180,8 @@ dependencies = [
     "onnx==1.20.1",
     "ipykernel>=7.2.0",
     "jupyter-core>=5.9.1",
+    "flashoptim>=0.1.3",
+    "black>=26.3.1",
 ]
 
 
