### 2026-03-30 c22b594
fix missing type hints and type errors
