+++ ncalab/training/distillation_trainer.py
@@ -1,336 +1,0 @@
-from __future__ import annotations
-import logging
-from pathlib import Path, PosixPath  # for type hint
-from typing import Dict, Optional, List, Tuple
-
-import numpy as np
-
-import torch  # type: ignore[import-untyped]
-import torch.optim as optim  # type: ignore[import-untyped]
-from torch.utils.data import DataLoader  # for type hint
-from torch.utils.tensorboard import SummaryWriter  # for type hint
-
-from tqdm import tqdm  # type: ignore[import-untyped]
-
-from ..models.basicNCA import BasicNCAModel  # for type hint
-from ..prediction import Prediction
-from ..utils import pad_input, unwrap
-from ..visualization import Visual
-
-from .earlystopping import EarlyStopping
-from .pool import Pool
-from .traininghistory import TrainingHistory
-
-
-class DistillationNCATrainer:
-    def __init__(
-        self,
-        nca: BasicNCAModel,
-        teacher: torch.nn.Module,
-        model_path: Path | PosixPath,
-        gradient_clipping: bool = False,
-        steps_range: tuple = (90, 110),
-        steps_validation: int = 100,
-        lr: Optional[float] = None,
-        lr_gamma: float = 0.99,
-        adam_betas=(0.4, 0.95),
-        batch_repeat: int = 2,
-        max_epochs: int = 200,
-        optimizer_method: str = "adamw",
-        pool: Optional[Pool] = None,
-    ):
-        """
-        :param nca: NCA model instance to train.
-        :type nca: ncalab.BasicNCAModel
-        :param model_path: Path to saved models. If None, models are not saved, defaults to None.
-        :type model_path: Path | PosixPath, optional
-        :param gradient_clipping: Whether to clip gradients, defaults to False.
-        :type gradient_clipping: bool, optional
-        :param steps_range: Inclusive range of NCA time steps, randomized in each forward pass, defaults to (90, 110).
-        :type steps_range: tuple, optional
-        :param steps_validation: Number of steps to use during validation, defaults to 100.
-        :type steps_validation: int, optional
-        :param lr: Initial learning rate, defaults to 16e-4.
-        :type lr: float, optional
-        :param lr_gamma: Exponential learning rate decay, defaults to 0.9999.
-        :type lr_gamma: float, optional
-        :param adam_betas: Beta values for Adam optimizer, defaults to (0.9, 0.95).
-        :type adam_betas: tuple, optional
-        :param batch_repeat: How often each batch will be duplicated, dfaults to 2.
-        :param max_epochs: Maximum number of epochs in training, defaults to 200.
-        :param optimizer_method: Optimization method, defaults to 'adam'.
-        :type optimizer_method: str, optional
-        :param pool: Sample pool object.
-        :type pool: ncalab.Pool
-        """
-        assert batch_repeat >= 1
-        assert steps_range[0] < steps_range[1]
-        assert max_epochs > 0
-        assert optimizer_method.lower() in (
-            "adam",
-            "adamw",
-            "adagrad",
-            "adafactor",
-            "rmsprop",
-            "sgd",
-        )
-        self.nca = nca
-        self.teacher = teacher
-        self.model_path = model_path
-        self.gradient_clipping = gradient_clipping
-        self.steps_range = steps_range
-        self.steps_validation = steps_validation
-        self.lr_gamma = lr_gamma
-        self.adam_betas = adam_betas
-        self.batch_repeat = batch_repeat
-        self.max_epochs = max_epochs
-        self.optimizer_method = optimizer_method
-        if lr is None:
-            if optimizer_method.lower() == "sgd":
-                self.lr = 1e-2
-            elif optimizer_method.lower() in ("adam", "adamw"):
[…]
