+++ pyproject.toml
@@ -96,7 +96,7 @@ dependencies = [
     "pyyaml==6.0.2",
     "pyzmq==26.4.0",
     "regex==2024.11.6",
-    "requests==2.32.4",
+    "requests==2.33.0",
     "rich==14.0.0",
     "safetensors==0.5.3",
     "scikit-image==0.25.2",
@@ -182,6 +182,7 @@ dependencies = [
     "jupyter-core>=5.9.1",
     "flashoptim>=0.1.3",
     "black>=26.3.1",
+    "cryptography==46.0.6",
 ]
 
 
