### 2026-03-30 abde869
update readme animations
