### 2026-03-30 5a5ad63
update medmnist examples to new validation api
