### 2026-03-30 665f27c
fix error in animator code
