### 2026-04-18 #9995 @schnorr
Regarding the "Development history and open-source practice", I have a
comment about the "Collaborative effort": looking to the commit
history with 447 commits, I can see 444 commits have been pushed by a
single person (@henrykrumb), and all Issues have been opened and
closed by the same person. So, while the article has six authors,
clearly there is one that contribute much more. I don't think this is
an issue as it seems all authors share the same lab and probably
exchange a lot of ideas that finally end up being carried out by the
same person.

Regarding the "Software paper", the checklist contains some required
sections that aren't present (at least not explicitely, although I've
tried to make the link with implicity content), namely: State of the
field (required section), Software Design (required section), AI usage
disclosure (required section). The Research impact statement I
implicit considered is section "Ongoing Research".

Finally, about the "General checks", I've left unmarked the
"Reproducibility" item because the project contains =flake.nix= and
=flake.lock= files indicating a possible very good effort of
instantiating a nix flake with a working software
stack. Unfortunetaly, I wasn't able to get it running. There is no
instruction about except the contents of the files and a simple commit
message a couple of months ago. It would be nice either to make it
working (I've removed the flake.lock so thax nix flake update would
work, but without much of a success), or remove it completely. A final
word about training. I've run the examples but since eval only works
after training, I couldn't afford the energy and time to wait for
training completition in my humble GPU. I would suggest authors to
include a very small scale problem that runs fast even on a CPU so
that any user can test the whole thing.

I'll wait for @henrykrumb to address my concerns.

For now, that's it @kanishkan91, sorry for the delay.
