### 2026-04-22 #9995 @sepandhaghighi
@henrykrumb

Firstly, I would like to express my appreciation for submitting this work to the JOSS journal.

My comments:

### Collaborative effort

I noticed that the repository has only one contributor. Furthermore, the `CONTRIBUTORS.md` file does not provide transparent information about each author's tasks.

### Community guidelines

The current version of the software lacks clarity on how new contributors can participate. It is essential to establish a clear contributing guideline, as well as **ISSUE** and **PR** templates, to facilitate a smooth and effective contribution process.

### Software design

The paper lacks a section dedicated to software design.

###  Dependencies

**NCALab** currently relies on over **170** primary dependencies, with some potentially being optional. This large number of dependencies may pose installation challenges in the future, and it is recommended to optimize this section.

### Test

Currently, **NCALab** includes automated tests that only run on the **Linux** operating system. I strongly recommend testing the package on other operating systems, such as **Windows** and **macOS**, and also on other stable versions of Python, including **Python 3.11**, **Python 3.13**, and **Python 3.14**. Additionally, please report the test coverage.

--------------

* I used [MyText v0.6](https://github.com/sepandhaghighi/mytext) (model: `meta/llama-3-8b-instruct`) to check the grammar in my review text.
* @kanishkan91 Sorry for my late response
