+++ pyproject.toml
@@ -47,7 +47,7 @@ dependencies = [
     "jedi==0.19.2",
     "jinja2==3.1.6",
     "joblib==1.4.2",
-    "keras==3.12.1",
+    "keras==3.13.2",
     "keras-applications==1.0.8",
     "kiwisolver==1.4.8",
     "lazy-loader==0.4",
@@ -76,7 +76,7 @@ dependencies = [
     "pandas-stubs==2.2.3.250308",
     "parso==0.8.4",
     "pexpect==4.9.0",
-    "pillow==12.1.1",
+    "pillow==12.2.0",
     "platformdirs==4.3.7",
     "pluggy==1.5.0",
     "pretrainedmodels==0.7.4",
@@ -89,7 +89,7 @@ dependencies = [
     "pydantic-core==2.33.2",
     "pygments==2.19.2",
     "pyparsing==3.2.3",
-    "pytest==8.3.5",
+    "pytest==9.0.3",
     "python-dateutil==2.9.0.post0",
     "pytorch-msssim==1.0.0",
     "pytz==2024.2",
@@ -125,7 +125,7 @@ dependencies = [
     "types-tqdm==4.67.0.20250417",
     "typing-extensions==4.13.2",
     "tzdata==2025.2",
-    "urllib3==2.6.3",
+    "urllib3==2.7.0",
     "wcwidth==0.2.13",
     "werkzeug==3.1.6",
     "wheel==0.46.2",
@@ -133,7 +133,7 @@ dependencies = [
     "zipp==3.21.0",
     "sphinx-autoapi>=3.6.0",
     "aiohappyeyeballs==2.6.1",
-    "aiohttp==3.13.3",
+    "aiohttp==3.13.4",
     "attrs==25.3.0",
     "build==1.2.2.post1",
     "cachetools==5.5.2",
@@ -169,7 +169,7 @@ dependencies = [
     "plyfile>=1.1",
     "twine>=6.1.0",
     "ruff>=0.12.4",
-    "torchmetrics>=1.7.4",
+    "torchmetrics==1.9.0",
     "segmentation-models-pytorch>=0.5.0",
     "scipy-stubs>=1.16.2.0",
     "fonttools>=4.57.0",
@@ -182,7 +182,7 @@ dependencies = [
     "jupyter-core>=5.9.1",
     "flashoptim>=0.1.3",
     "black>=26.3.1",
-    "cryptography==46.0.6",
+    "cryptography==46.0.7",
 ]
 
 
