### 2026-05-13 384904c
add script to generate and save animations
