### 2026-05-13 2ca9a43
add animators to evaluation scripts
