+++ pyproject.toml
@@ -87,7 +87,7 @@ dependencies = [
     "pure-eval==0.2.3",
     "pydantic==2.11.4",
     "pydantic-core==2.33.2",
-    "pygments==2.19.2",
+    "pygments==2.20.0",
     "pyparsing==3.2.3",
     "pytest==9.0.3",
     "python-dateutil==2.9.0.post0",
@@ -119,7 +119,6 @@ dependencies = [
     "tokenizers==0.21.1",
     "torch-tb-profiler==0.4.3",
     "traitlets==5.14.3",
-    "transformers==4.53",
     "types-pytz==2025.2.0.20250326",
     "types-requests==2.32.0.20250328",
     "types-tqdm==4.67.0.20250417",
@@ -176,13 +175,14 @@ dependencies = [
     "torch==2.8.0",
     "tensorboard-plugin-profile>=2.21.6",
     "tornado>=6.5.5",
-    "uv>=0.10.11",
+    "uv==0.11.6",
     "onnx==1.20.1",
     "ipykernel>=7.2.0",
     "jupyter-core>=5.9.1",
     "flashoptim>=0.1.3",
     "black>=26.3.1",
     "cryptography==46.0.7",
+    "pip==26.1",
 ]
 
 
