### 2026-05-13 4152112
add image gradient based animation for segmentations
