### 2026-05-13 0db00b3
update segmentation visual in tensorboard
