+++ pyproject.toml
@@ -10,180 +10,171 @@ classifiers = [
     "Development Status :: 3 - Alpha"
 ]
 dependencies = [
-    "absl-py==2.2.2",
+    "absl-py==2.4.0",
     "albucore==0.0.24",
     "albumentations==2.0.8",
     "annotated-types==0.7.0",
     "asttokens==3.0.1",
     "astunparse==1.6.3",
     "certifi==2025.4.26",
-    "charset-normalizer==3.4.2",
-    "click==8.1.8",
-    "comm==0.2.2",
-    "contourpy==1.3.2",
+    "charset-normalizer==3.4.7",
+    "click==8.3.3",
+    "comm==0.2.3",
+    "contourpy==1.3.3",
     "cycler==0.12.1",
-    "debugpy==1.8.14",
+    "debugpy==1.8.20",
     "decorator==5.2.1",
-    "numpy==2.1.3",
-    "tqdm==4.67.1",
-    "emoji==2.14.1",
-    "etils==1.12.2",
-    "eval-type-backport==0.2.2",
-    "executing==2.2.0",
-    "filelock==3.20.3",
-    "fire==0.7.0",
-    "flatbuffers==25.2.10",
+    "numpy==2.4.4",
+    "tqdm==4.67.3",
+    "executing==2.2.1",
+    "filelock==3.24.2",
+    "fire==0.7.1",
+    "flatbuffers==25.12.19",
     "fsspec==2025.3.2",
-    "gast==0.6.0",
+    "gast==0.7.0",
     "google-pasta==0.2.0",
-    "grpcio==1.72.0",
+    "grpcio==1.80.0",
     "gviz-api==1.10.0",
     "h5py==3.12.1",
-    "idna==3.10",
-    "imageio==2.37.0",
-    "importlib-resources==6.5.2",
-    "iniconfig==2.1.0",
-    "ipython==8.31.0",
-    "jedi==0.19.2",
+    "idna==3.15",
+    "imageio==2.37.3",
+    "importlib-resources==7.1.0",
+    "iniconfig==2.3.0",
+    "ipython==9.13.0",
+    "jedi==0.20.0",
     "jinja2==3.1.6",
-    "joblib==1.4.2",
-    "keras==3.13.2",
+    "joblib==1.5.3",
+    "keras==3.14.1",
     "keras-applications==1.0.8",
-    "kiwisolver==1.4.8",
-    "lazy-loader==0.4",
+    "kiwisolver==1.5.0",
+    "lazy-loader==0.5",
     "libclang==18.1.1",
-    "markdown==3.8.1",
-    "markdown-it-py==3.0.0",
-    "markupsafe==3.0.2",
-    "matplotlib==3.10.1",
-    "matplotlib-inline==0.1.7",
+    "markdown==3.10.2",
+    "markdown-it-py==4.2.0",
+    "markupsafe==3.0.3",
+    "matplotlib==3.10.9",
+    "matplotlib-inline==0.2.2",
     "mdurl==0.1.2",
     "medmnist==3.0.2",
-    "ml-dtypes==0.5.1",
+    "ml-dtypes==0.5.4",
     "mpmath==1.3.0",
     "munch==4.0.0",
-    "mypy==1.15.0",
+    "mypy==2.1.0",
     "mypy-extensions==1.1.0",
-    "namex==0.0.9",
+    "namex==0.1.0",
     "nest-asyncio==1.6.0",
-    "networkx==3.4.2",
-    "onnxscript==0.2.5",
-    "opencv-python-headless==4.11.0.86",
+    "networkx==3.6.1",
+    "onnxscript==0.7.0",
[…]
