+++ pyproject.toml
@@ -156,7 +156,7 @@ dependencies = [
     "scipy-stubs==1.17.1.4",
     "fonttools==4.62.1",
     "tornado>=6.5.5",
-    "uv==0.11.14",
+    "uv==0.11.15",
     "onnx==1.21.0",
     "ipykernel>=7.2.0",
     "jupyter-core>=5.9.1",
