### 2026-06-01 0ed834f
fix type errors in trainer
