+++ .pre-commit-config.yaml
@@ -10,12 +10,12 @@ repos:
     -   id: check-added-large-files
         args: [--maxkb=4096]
 -   repo: https://github.com/pre-commit/mirrors-mypy
-    rev: 'v1.19.1'
+    rev: 'v2.1.0'
     hooks:
     -   id: mypy
 -   repo: https://github.com/astral-sh/ruff-pre-commit
     # Ruff version.
-    rev: v0.15.5
+    rev: v0.15.15
     hooks:
     # Run the linter.
     -   id: ruff-check
