### 2026-06-01 9bb3301
add validation batch visual to tensorboard
