### 2026-06-15 ffe20c1
use composite loss in segmentation NCA
