+++ pyproject.toml
@@ -123,7 +123,7 @@ dependencies = [
     "zipp==3.23.1",
     "sphinx-autoapi==3.8.0",
     "aiohappyeyeballs==2.6.1",
-    "aiohttp==3.13.5",
+    "aiohttp==3.14.0",
     "attrs==26.1.0",
     "build==1.5.0",
     "frozenlist==1.8.0",
@@ -155,7 +155,7 @@ dependencies = [
     "segmentation-models-pytorch>=0.5.0",
     "scipy-stubs==1.17.1.4",
     "fonttools==4.62.1",
-    "tornado>=6.5.5",
+    "tornado==6.5.6",
     "uv==0.11.15",
     "onnx==1.21.0",
     "ipykernel>=7.2.0",
@@ -174,7 +174,7 @@ dependencies = [
     "nh3==0.3.5",
     "huggingface-hub==1.14.0",
     "pipdeptree>=2.35.2",
-    "torch==2.11.0",
+    "torch==2.12.0",
 ]
 
 
