### 2026-06-16 35db365
significantly improve prediction API: segmentation masks, convenience functions
