### 2026-06-16 e6a7270
refactor animator API
- make use of new prediction API features (unwrapping over time)
- start simplifying code
- externalize modules for color handling and visualization
