### 2026-06-16 6bf7fe6
streamline application-specific models
- detach() loss tensors used for logging
- remove num_classess setter in classification
- rename SegmentationNCAModel --> BinarySegmentationNCAModel
- fix bad handling of logits and final predictions in segmentation
