### 2026-06-16 e47feed
adapt kvasir-seg segmentation task to new API
