### 2026-06-16 5f273a2
fix tests for growing model, add tests for prediction API
