### 2026-06-16 aac3d80
bump dependencies, remove stale dependencies
