+++ pyproject.toml
@@ -7,174 +7,23 @@ requires-python = ">=3.12"
 license = "MIT"
 classifiers = [
     "Programming Language :: Python :: 3.12",
-    "Development Status :: 3 - Alpha"
+    "Development Status :: 3 - Alpha",
 ]
 dependencies = [
-    "absl-py==2.4.0",
-    "albucore==0.0.24",
-    "albumentations==2.0.8",
-    "annotated-types==0.7.0",
-    "asttokens==3.0.1",
-    "astunparse==1.6.3",
-    "certifi==2025.4.26",
-    "charset-normalizer==3.4.7",
-    "click==8.3.3",
-    "comm==0.2.3",
-    "contourpy==1.3.3",
-    "cycler==0.12.1",
-    "debugpy==1.8.20",
-    "decorator==5.2.1",
-    "numpy==2.4.4",
-    "tqdm==4.67.3",
-    "executing==2.2.1",
-    "filelock==3.24.2",
-    "fire==0.7.1",
-    "flatbuffers==25.12.19",
-    "fsspec==2025.3.2",
-    "gast==0.7.0",
-    "google-pasta==0.2.0",
-    "grpcio==1.80.0",
-    "gviz-api==1.10.0",
-    "h5py==3.12.1",
-    "idna==3.15",
-    "imageio==2.37.3",
-    "importlib-resources==7.1.0",
-    "iniconfig==2.3.0",
-    "ipython==9.13.0",
-    "jedi==0.20.0",
-    "jinja2==3.1.6",
-    "joblib==1.5.3",
-    "keras==3.14.1",
-    "keras-applications==1.0.8",
-    "kiwisolver==1.5.0",
-    "lazy-loader==0.5",
-    "libclang==18.1.1",
-    "markdown==3.10.2",
-    "markdown-it-py==4.2.0",
-    "markupsafe==3.0.3",
-    "matplotlib==3.10.9",
-    "matplotlib-inline==0.2.2",
-    "mdurl==0.1.2",
-    "medmnist==3.0.2",
-    "ml-dtypes==0.5.4",
-    "mpmath==1.3.0",
-    "munch==4.0.0",
-    "mypy==2.1.0",
-    "mypy-extensions==1.1.0",
-    "namex==0.1.0",
-    "nest-asyncio==1.6.0",
-    "networkx==3.6.1",
-    "onnxscript==0.7.0",
-    "opencv-python-headless==4.13.0.92",
-    "opt-einsum==3.4.0",
-    "optree==0.19.1",
-    "packaging==26.2",
-    "pandas==3.0.3",
-    "pandas-stubs==3.0.0.260204",
-    "parso==0.8.7",
-    "pexpect==4.9.0",
-    "pillow==12.2.0",
-    "platformdirs==4.9.6",
-    "pluggy==1.6.0",
-    "pretrainedmodels==0.7.4",
-    "prompt-toolkit==3.0.52",
-    "protobuf==7.34.1",
-    "psutil==7.2.2",
-    "ptyprocess==0.7.0",
-    "pure-eval==0.2.3",
-    "pydantic==2.13.4",
-    "pydantic-core==2.46.4",
-    "pygments==2.20.0",
-    "pyparsing==3.3.2",
-    "pytest==9.0.3",
-    "python-dateutil==2.9.0.post0",
-    "pytorch-msssim==1.0.0",
-    "pyyaml==6.0.3",
-    "pyzmq==27.1.0",
-    "requests==2.34.0",
-    "rich==15.0.0",
-    "safetensors==0.7.0",
-    "scikit-image==0.26.0",
[…]
