### 2026-06-16 5e8c64f
segmentation example: add early stopping
