### 2026-06-16 31acb0e
update growing emoji example to API
