### 2026-06-16 eda02f8
fix missing doc dependencies
