+++ pyproject.toml
@@ -19,6 +19,8 @@ dependencies = [
     "pytorch-msssim>=1.0.0",
     "requests>=2.34.0",
     "safetensors>=0.8.0",
+    "sphinx-autoapi>=3.8.0",
+    "sphinx-rtd-theme>=3.1.0",
     "tensorboard>=2.20.0",
     "torch>=2.12.0",
     "torchmetrics>=1.9.0",
