### 2026-06-16 f70e3d9
fix missing dependencies for quality control
