+++ pyproject.toml
@@ -14,10 +14,13 @@ dependencies = [
     "click>=8.4.1",
     "matplotlib>=3.11.0",
     "medmnist>=3.0.2",
+    "mypy>=2.1.0",
     "numpy>=2.4.6",
     "opencv-python>=4.13.0.92",
+    "pytest>=9.1.0",
     "pytorch-msssim>=1.0.0",
     "requests>=2.34.0",
+    "ruff>=0.15.17",
     "safetensors>=0.8.0",
     "sphinx-autoapi>=3.8.0",
     "sphinx-rtd-theme>=3.1.0",
