### 2026-06-16 915ac9d
bump ruff pre-commit hook
