+++ .pre-commit-config.yaml
@@ -15,7 +15,7 @@ repos:
     -   id: mypy
 -   repo: https://github.com/astral-sh/ruff-pre-commit
     # Ruff version.
-    rev: v0.15.15
+    rev: v0.15.17
     hooks:
     # Run the linter.
     -   id: ruff-check
