### 2026-06-16 9a13a26
add type stub dependencies
