+++ pyproject.toml
@@ -17,11 +17,13 @@ dependencies = [
     "mypy>=2.1.0",
     "numpy>=2.4.6",
     "opencv-python>=4.13.0.92",
+    "pandas-stubs>=3.0.3.260530",
     "pytest>=9.1.0",
     "pytorch-msssim>=1.0.0",
     "requests>=2.34.0",
     "ruff>=0.15.17",
     "safetensors>=0.8.0",
+    "scipy-stubs>=1.17.1.5",
     "sphinx-autoapi>=3.8.0",
     "sphinx-rtd-theme>=3.1.0",
     "tensorboard>=2.20.0",
