+++ CONTRIBUTORS.md
@@ -1,5 +1,5 @@
-* Henry Krumb, TU Darmstadt
-* Richard Sattel, TU Darmstadt
-* Dennis Grotz, TU Darmstadt
-* Jonathan Dewenter, TU Darmstadt
-* Niklas Ihm, TU Darmstadt
+# List of contributors and their contributions to NCALab
+
+* **Henry Krumb**, TU Darmstadt - Project maintainer and initiator
+* **Richard Sattel**, TU Darmstadt - Provided NixOS config and feedback on the codebase
+* **Anirban Mukhopadhyay**, TU Darmstadt - Head of research lab, providing resources and advice
