+++ .github/ISSUE_TEMPLATE/question.md
@@ -1,0 +1,10 @@
+---
+name: Question
+about: Raise questions about the scope of this project or implementation details
+title: "[QUESTION] "
+labels: question
+assignees: ''
+
+---
+
+
