### 2026-06-17 1b3e8d3
include python 3.11, 3.13 and 3.14 in package workflow
