+++ .github/workflows/python-package.yml
@@ -19,13 +19,13 @@ jobs:
         python-version: ["3.10", "3.11", "3.12", "3.13", "3.14"]
 
     steps:
-    - uses: actions/checkout@v3
+    - uses: actions/checkout@v6
     - name: Install uv
       uses: astral-sh/setup-uv@v6
       with:
-          version: "0.8.2"
+          version: "0.11.14"
     - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v3
+      uses: actions/setup-python@v6
       with:
         python-version: ${{ matrix.python-version }}
     - name: Lint with ruff
