### 2026-06-17 58c8d36
group dependencies into 'docs' and 'dev'
