+++ pyproject.toml
@@ -14,18 +14,13 @@ dependencies = [
     "click>=8.4.1",
     "matplotlib>=3.11.0",
     "medmnist>=3.0.2",
-    "mypy>=2.1.0",
     "numpy>=2.4.6",
     "opencv-python>=4.13.0.92",
     "pandas-stubs>=3.0.3.260530",
-    "pytest>=9.1.0",
     "pytorch-msssim>=1.0.0",
     "requests>=2.34.0",
-    "ruff>=0.15.17",
     "safetensors>=0.8.0",
     "scipy-stubs>=1.17.1.5",
-    "sphinx-autoapi>=3.8.0",
-    "sphinx-rtd-theme>=3.1.0",
     "tensorboard>=2.20.0",
     "torch>=2.12.0",
     "torchmetrics>=1.9.0",
@@ -43,8 +38,18 @@ Repository = "https://github.com/MECLabTUDA/NCAlab"
 packages = ["ncalab"]
 
 [dependency-groups]
-docs = ["sphinx-autoapi>=3.6.0", "sphinx>=8.1.3", "sphinx-rtd-theme>=3.0.2"]
-doc = ["sphinx>=8.1.3", "sphinx-rtd-theme>=3.0.2"]
+dev = [
+    "black>=26.5.1",
+    "coverage>=7.14.1",
+    "mypy>=2.1.0",
+    "pytest>=9.1.0",
+    "ruff>=0.15.17",
+]
+docs = [
+    "sphinx>=9.1.0",
+    "sphinx-autoapi>=3.6.0",
+    "sphinx-rtd-theme>=3.0.2",
+]
 
 
 [tool.uv.sources]
