+++ .github/workflows/docs.yml
@@ -24,6 +24,7 @@ jobs:
       uses: actions/setup-python@v6
       with:
         python-version: ${{ matrix.python-version }}
-    - name: Install dependencies
+    - name: Build docs
       run: |
+        uv sync --group docs
         uv run sphinx-build -b html docs/source/ docs/build/html
