### 2026-06-17 506c71c
update gif animations for use in readme
