### 2026-06-17 a0db471
adapt examples to changes in API
