### 2026-06-18 4864b55
update segmentation animation in readme
