### 2026-06-18 b41f8b8
fix selfclassifying MNIST example, split selfclassification from classification model class
