### 2026-06-18 eb12a92
fix missing use_laplace parameter
