### 2026-06-18 5d8e65c
add first draft of contribution guidelines
