+++ CONTRIBUTING.md
@@ -1,0 +1,39 @@
+# Introduction
+
+Thank you for your interest in contributing to NCALab.
+Free Open Source software benefits from any kind of support and even small contributions can make a large difference.
+
+We are open to contributions, including but not limited to:
+* Contributions of new NCA-type architectures
+* Enhancements that simplify the creation, training or evaluation of NCAs
+* Performance improvements
+* New example tasks
+* Bug fixes
+* Documentation
+
+In the following sections, you will find some guidelines on how to participate.
+
+
+## Development
+
+Please find a set of useful commands in the project readme under "for developers".
+
+
+## Issue Policy
+
+You have found a bug, a missing feature or have a question about NCALab itself?
+Then please do not hesitate to write an issue on GitHub.
+There are only few things we would appreciate if they are considered:
+* Briefly search existing issues before creating a new one. If an issue exists, you may add details in the comment section.
+* If your issue describes a bug, we appreciate a detailed description of the symptoms and instructions to reproduce.
+* Please use one of the issue templates provided in this repo, if applicable.
+
+
+## Pull Request Policy
+
+If your PR was merged successfully, you will be asked for your consent to be added to the contributor's list.
+
+Example tasks should be as concise as possible without too much extra overhead.
+Please use publicly available datasets (licensed Creative Commons or similar) with proper attribution if required.
+
+Contributions solely made by AI agents or other automated systems are rejected by default.
