### 2026-06-19 6e25e0f
add readme badge for code coverage
