+++ pyproject.toml
@@ -22,6 +22,7 @@ dependencies = [
     "requests>=2.34.0",
     "safetensors>=0.8.0",
     "scipy-stubs>=1.17.1.5",
+    "tabulate>=0.10.0",
     "tensorboard>=2.20.0",
     "torch>=2.12.0",
     "torchmetrics>=1.9.0",
