### 2026-06-19 e9f41b0
github workflow: add coverage badge
