### 2026-06-19 75b432a
add test cases, fix typo, remove stale code
