+++ ncalab/models/wrappers/ensemble.py
@@ -1,3 +1,0 @@
-class EnsembleNCA:
-    def __init__(self, ncas):
-        self.ncas = ncas
