+++ ncalab/visualization/confusionmatrix.py
@@ -1,12 +1,0 @@
-class ConfusionMatrix:
-    def __init__(self):
-        pass
-
-    def generate_image(self):
-        pass
-
-    def to_string(self) -> str:
-        return ""
-
-    def __str__(self):
-        return self.to_string()
