### 2026-06-19 16a0ea6
update pre commit config
