### 2026-06-19 94a501e
remove coverage badge generation from github workflow, now handled in pre-commit
