+++ .github/workflows/python-package.yml
@@ -36,11 +36,7 @@ jobs:
     - name: Test with pytest
       run: |
         uv sync --group dev
-        uv run coverage run --source ncalab -m pytest tests
-        uv run coverage xml
-    - name: Create coverage badge
-      run: |
-        uv run python -m scripts/coveragebadge.py
+        uv run pytest tests
     - name: Type checking with mypy
       run: |
         uv sync --group dev
