### 2026-06-19 03a5ddb
update dev instructions in readme
