### 2026-06-19 9a23b3d
segmentation example: update animation, compile torch model
