### 2026-06-19 7db3339
medmnist example: refactor code
