### 2026-06-19 bd7b0ec
weights download script: dynamically load project version
