### 2026-06-19 524bf6f
save animations script: run without GPU
