### 2026-06-19 0dc696a
add comments and command line args to scripts
