### 2026-06-19 be9405c
add unit tests for utils and autostepper
