### 2026-06-19 53d3e77
fix finetuning script for growing emoji example
