### 2026-06-19 b7e6622
script for packing example weights: show file sizes
