### 2026-06-19 3614267
add comments and function stubs
