### 2026-06-19 92d8dc1
update checksum in download_example_weights script
