### 2026-06-19 ada041f
finalize changelog
