### 2026-06-29 7fda32b
animator: improve animatorstyle capabilities
