### 2026-06-29 39d2ab7
bump dependency versions
