+++ pyproject.toml
@@ -11,23 +11,30 @@ classifiers = [
 ]
 dependencies = [
     "albumentations>=2.0.8",
-    "click>=8.4.1",
+    "certifi==2026.6.17",
+    "click==8.4.2",
+    "coverage==7.14.3",
     "cuda-toolkit>=13.2.1",
     "matplotlib>=3.11.0",
     "medmnist>=3.0.2",
     "numpy>=2.4.6",
+    "numpy-typing-compat==20260602.2.5",
     "opencv-python>=4.13.0.92",
+    "optype[numpy]==0.18.0",
     "pandas-stubs>=3.0.3.260530",
+    "pytest==9.1.1",
     "pytorch-msssim>=1.0.0",
-    "requests>=2.34.0",
+    "requests==2.34.2",
+    "ruff==0.15.20",
     "safetensors>=0.8.0",
-    "scipy-stubs>=1.17.1.5",
+    "scipy==1.18.0",
+    "scipy-stubs==1.18",
     "tabulate>=0.10.0",
     "tensorboard>=2.20.0",
-    "torch>=2.12.0",
+    "torch==2.12.1+cu132",
     "torchmetrics>=1.9.0",
     "torchvision>=0.27.0",
-    "tqdm>=4.68.2",
+    "tqdm==4.68.3",
 ]
 
 
