### 2026-06-29 a1fe521
remove badge generation from pre-commit config
