### 2026-06-29 c6120d0
coverage script: run pytest
