### 2026-06-29 efbc351
remove stale file
