### 2026-06-29 e1a9164
write additional unit tests for earlystopping, autostepper, utils and losses
