### 2026-06-29 377256f
improve docs of growing emoji task
