### 2026-06-29 da586f5
remove optimizer init bloat from trainer
