+++ ncalab/training/trainer.py
@@ -30,14 +30,12 @@ class BasicNCATrainer:
         nca: AbstractNCAModel,
         model_path: Optional[Path | PosixPath],
         gradient_clipping: bool = False,
-        lr: Optional[float] = None,
-        lr_gamma: float = 0.99,
-        adam_betas=(0.9, 0.95),
+        optimizer: Optional[optim.Optimizer] = None,
+        lr: float = 1e-3,
+        lr_scheduler: Optional[optim.lr_scheduler.LRScheduler] = None,
         batch_repeat: int = 2,
         max_epochs: int = 200,
-        optimizer_method: str = "adam",
         pool: Optional[Pool] = None,
-        lr_scheduler: Optional[optim.lr_scheduler.LRScheduler] = None,
     ):
         """
         :param nca: NCA model instance to train.
@@ -48,48 +46,25 @@ class BasicNCATrainer:
         :type gradient_clipping: bool, optional
         :param lr: Initial learning rate, defaults to 16e-4.
         :type lr: float, optional
-        :param lr_gamma: Exponential learning rate decay, defaults to 0.9999.
-        :type lr_gamma: float, optional
-        :param adam_betas: Beta values for Adam optimizer, defaults to (0.9, 0.95).
-        :type adam_betas: tuple, optional
         :param batch_repeat: How often each batch will be duplicated, dfaults to 2.
         :param max_epochs: Maximum number of epochs in training, defaults to 200.
-        :param optimizer_method: Optimization method, defaults to 'adam'.
-        :type optimizer_method: str, optional
         :param pool: Sample pool object.
         :type pool: ncalab.Pool
         """
         assert batch_repeat >= 1
         assert max_epochs > 0
-        assert optimizer_method.lower() in (
-            "adam",
-            "adamw",
-            "adagrad",
-            "adafactor",
-            "rmsprop",
-            "sgd",
-        )
+
         self.nca = nca
         self.model_path = model_path
         self.gradient_clipping = gradient_clipping
-        self.lr_gamma = lr_gamma
-        self.adam_betas = adam_betas
         self.batch_repeat = batch_repeat
         self.max_epochs = max_epochs
-        self.optimizer_method = optimizer_method
-        if lr is None:
-            if optimizer_method.lower() == "sgd":
-                self.lr = 1e-2
-            elif optimizer_method.lower() in ("adam", "adamw"):
-                self.lr = 1e-3
-            elif optimizer_method.lower() == "rmsprop":
-                self.lr = 1e-2
-            elif optimizer_method.lower() == "adagrad":
-                self.lr = 1e-2
-            else:
-                self.lr = 1e-2
-        else:
-            self.lr = lr
+        self.lr = lr
+        self.optimizer = (
+            optim.Adam(self.nca.parameters(), self.lr, betas=(0.9, 0.95))
+            if optimizer is None
+            else optimizer
+        )
         self.pool = pool
         self.lr_scheduler = lr_scheduler
 
@@ -98,19 +73,16 @@ class BasicNCATrainer:
         Shows a markdown-formatted info string with training parameters.
         Useful for showing info on tensorboard to keep track of parameter changes.
 
-        :returns [str]: Markdown-formatted info string.
+        :return: Markdown-formatted info string.
+        :rtype: str
         """
         s = "BasicNCATrainer Info\n"
         s += "-------------------\n"
         for attribute in (
             "model_path",
-            "lr",
-            "lr_gamma",
             "gradient_clipping",
-            "adam_betas",
             "batch_repeat",
             "max_epochs",
-            "optimizer_method",
         ):
             attribute_f = attribute.title().replace("_", " ")
[…]
