### 2026-06-29 50ee5e8
update whitepaper: section on software design
