### 2026-06-29 a396f7f
improve type hints in uncertainty estimator
