### 2026-06-29 3fa65c6
medmnist task: adjust default settings
