### 2026-06-29 5175cb7
add acknowledgements to whitepaper
