### 2026-06-29 45acb69
sample pool: revise docs
