### 2026-06-29 98106b9
indicate supported python versions in README
