+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "ncalab"
-version = "0.5.0"
+version = "0.5.1"
 description = "Training code and pre-trained models for Neural Cellular Automata on different datasets and tasks"
 readme = "README.md"
 requires-python = ">=3.12"
