### 2026-08-17 #9995 @henrykrumb
Dear @sepandhaghighi ,

thank you very much for the updated review. We addressed your comments and respond to them in the following:

1. Thank you for pointing us to this. We have added a pull request template in `.github/PULL_REQUEST_TEMPLATE/default.md`.
2. We are working continuously to improve the test coverage. In the most recent release, we reach a coverage of 70.46%, which is already a slight improvement compared to the 66% reported earlier. We are planning to reach at least 75% test coverage in the near future.
3. Wonderful, thank you very much for this assessment.
4. We have revised the `Software Design` section, added more contents and inserted a block diagram to illustrate the core of our project (canonical NCA architecture).

Best regards
Henry Krumb
